+++ src/sourceTracking.ts
@@ -29,7 +29,10 @@ import {
 // this is not exported by SDR (see the comments in SDR regarding its limitations)
 import { filePathsFromMetadataComponent } from '@salesforce/source-deploy-retrieve/lib/src/utils/filePathGenerator';
 import { Performance } from '@oclif/core/performance';
-import { RemoteSourceTrackingService, remoteChangeElementToChangeResult } from './shared/remoteSourceTrackingService';
+import {
+  RemoteSourceTrackingService,
+  remoteChangeElementToChangeResult,
+} from './shared/remote/remoteSourceTrackingService';
 import { ShadowRepo } from './shared/local/localShadowRepo';
 import { throwIfConflicts, findConflictsInComponentSet, getDedupedConflictsFromChanges } from './shared/conflicts';
 import {
@@ -327,7 +330,7 @@ export class SourceTracking extends AsyncCreatable {
       if (options.format === 'ChangeResultWithPaths') {
         return populateFilePaths({
           elements: filteredChanges.map(remoteChangeElementToChangeResult),
-          packageDirPaths: this.project.getPackageDirectories().map((pkgDir) => pkgDir.path),
+          packageDirPaths: this.project.getPackageDirectories().map((pkgDir) => pkgDir.fullPath),
           registry: this.registry,
         });
       }