+++ test/unit/localDetectMovedFiles.test.ts
@@ -10,7 +10,7 @@ import * as fs from 'node:fs';
 // @ts-expect-error isogit has both ESM and CJS exports but node16 module/resolution identifies it as ESM
 import git from 'isomorphic-git';
 import { expect, config } from 'chai';
-import sinon = require('sinon');
+import sinon from 'sinon';
 import { RegistryAccess } from '@salesforce/source-deploy-retrieve';
 import { ShadowRepo } from '../../src/shared/local/localShadowRepo';